Step 4 client side done, passing to Ron.

+++ java/tenmo-client/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -3,23 +3,23 @@ package com.techelevator.tenmo.model;
 import java.math.BigDecimal;
 
 public class Account {
-    private Long accountId;
-    private Long userId;
+    private int accountId;
+    private int userId;
     private BigDecimal balance;
 
-    public Long getAccountId() {
+    public int getAccountId() {
         return accountId;
     }
 
-    public void setAccountId(Long accountId) {
+    public void setAccountId(int accountId) {
         this.accountId = accountId;
     }
 
-    public Long getUserId() {
+    public int getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(int userId) {
         this.userId = userId;
     }
